Pilot spec 0005 on BrandResource
Extract the three real Lunar concerns inline on BaseResource into named
traits under Support/Resources/Concerns/:

  - HasLunarPermissions: $permission, hasPermission(), can(), and the
    registerNavigationItems() override.
  - ResolvesModelContract: the interface-aware getModel() override.
  - HasScoutGlobalSearch: applyGlobalSearchAttributeConstraints() plus
    mapSearchableAttributes() for Scout-enabled global search.

BaseResource now mirrors Filament's own Resource.php shape — thin base
that composes named concerns. The wrapper traits (ExtendsForms,
ExtendsTables, ExtendsPages, ExtendsRelationManagers, ExtendsSubnavigation)
remain for resources still on the legacy shape and are marked
@deprecated; they are removed once every resource is migrated.

Pilot migration: BrandResource adopts the v5 split-class layout:

  BrandResource/
  ├── Schemas/BrandForm.php   ← form components + granular getters
  ├── Tables/BrandTable.php   ← table columns + actions
  └── Pages/*                 (unchanged)

BrandResource.php drops to just metadata + thin form()/table()
delegation to the split classes. Hook target moves from
BrandResource::class to BrandForm::class / BrandTable::class with method
names configureForm/configureTable, matching the spec's resolved
extension model.
