+++ packages/admin/src/Filament/Resources/BrandResource/Tables/BrandTable.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace Lunar\Admin\Filament\Resources\BrandResource\Tables;
+
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Tables\Columns\SpatieMediaLibraryImageColumn;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Table;
+use Lunar\Admin\Support\Concerns\CallsHooks;
+
+class BrandTable
+{
+    use CallsHooks;
+
+    public static function configure(Table $table): Table
+    {
+        return self::callStaticLunarHook(
+            'configureTable',
+            $table
+                ->columns(static::getColumns())
+                ->filters([])
+                ->recordActions([
+                    EditAction::make(),
+                ])
+                ->toolbarActions([
+                    BulkActionGroup::make([
+                        DeleteBulkAction::make(),
+                    ]),
+                ])
+                ->searchable(),
+        );
+    }
+
+    public static function getColumns(): array
+    {
+        return [
+            SpatieMediaLibraryImageColumn::make('thumbnail')
+                ->collection(config('lunar.media.collection'))
+                ->conversion('small')
+                ->limit(1)
+                ->square()
+                ->label(''),
+            TextColumn::make('name')
+                ->label(__('lunarpanel::brand.table.name.label'))
+                ->searchable(),
+            TextColumn::make('products_count')
+                ->counts('products')
+                ->formatStateUsing(
+                    fn ($state) => number_format($state, 0)
+                )
+                ->label(__('lunarpanel::brand.table.products_count.label')),
+        ];
+    }
+}
