+++ packages/admin/src/Support/Resources/Concerns/ResolvesModelContract.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Lunar\Admin\Support\Resources\Concerns;
+
+use ReflectionClass;
+
+trait ResolvesModelContract
+{
+    public static function getModel(): string
+    {
+        $class = new ReflectionClass(static::$model);
+
+        if ($class->isInterface()) {
+            return app()->get(static::$model)::class;
+        }
+
+        return parent::getModel();
+    }
+}
